requirements.txt
flask==2.3.3
flask-cors==4.0.0
pandas==2.2.3
openpyxl==3.1.2
pillow==10.0.1
weasyprint==60.1
python-dotenv==1.0.0